Add ignoreExistingCoverage true
The project has a couple of tests that cover ~10% of the code. To generate 100% coverage for Java Demo on dashboard, we need this option. Previously, it was not needed to have this option explicitly set in this file because dashboard was setting it by default. This is no longer true due to platform API change.